chore(app): weatherview 컴포넌트 제거 및 weather 위젯으로 통합, DateTime 컴포넌트에 요일 표시 기능 추가, 반복되는 날씨 정보 map으로 리팩토링
