--- apps/yongin-platform-app/src/components/WeatherView/index.ts
@@ -1,1 +1,0 @@
-export { WeatherView } from "./WeatherView";
